stp_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0
numpy==1.24.3
python-multipart==0.0.6
spacy==3.7.2
https://github.com/explosion/spacy-models/releases/download/tr_core_news_sm-3.7.0/tr_core_news_sm-3.7.0-py3-none-any.whl
openai==1.3.8 